\name{analogues-package}
\alias{analogues-package}
\alias{analogues}
\docType{package}
\title{
Identification of climate analogues
}
\description{
Uses a climatic similarity statistic (based on the Euclidean distance) in conjunction with WorldClim and downscaled Global Climate Model (GCM) data to connect areas with similarity climates which may be spatially and/or temporally separated. A GUI is available to facilitate the more commonly used search functions.
Depends: raster, rgdal, sp, stringr.
}
\details{
\tabular{ll}{
Package: \tab Analogues\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2020-01-29\cr
License: \tab GPL 2\cr
}
Select your location, search area, variables, growing season, rotation, and climate data. Then run to identify areas with similar climatic condition.
}
\author{
David Arango Lodono, Edward Jones, and Julian Ramirez-Villegas

Maintainer: Julian Ramirez-Villegas <j.r.villegas@cgiar.org>
}
\references{
Details of the CCAFS Climate Analogues technique in general may be found here:
Ramirez-Villegas J, Lau C, Koehler A-K, Signer J, Jarvis A, Arnell N, Osborne T, Hooker J. 2011. Climate analogues: finding tomorrow's agriculture today. Working Paper no. 12. Cali, Colombia: CGIAR.
}
\keyword{ package }
