\name{calcCcafs.matrix}
\alias{calcCcafs.matrix}
\title{Eucledian distance to calculate dissimilarity...}
\usage{\method{calcCcafs}{matrix} (lag, params, training, weights, base,
    project)\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
\method{calcCcafs}{matrix}(lag, params, training, weights, base, project)
}
\description{Eucledian distance to calculate dissimilarity}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

\alias{calcCcafs.matrix}

