\name{calcCcafs}
\alias{calcCcafs}
\title{Eucledian distance to calculate dissimilarity...}
\usage{calcCcafs(lag, params, training, weights, base, project)calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
calcCcafs(lag, params, training, weights, base, project)
}
\description{Eucledian distance to calculate dissimilarity}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

\alias{calcCcafs}

