\name{calcDis}
\alias{calcDis}
\title{Call the right funciton to calculate distance...}
\usage{calcDis(roll, params, training, weights, base, project)calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
calcDis(roll, params, training, weights, base, project)
}
\description{Call the right funciton to calculate distance}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

\alias{calcDis}

