\name{calcHal.matrix}
\alias{calcHal.matrix}
\title{Method after Hallegatte...}
\usage{\method{calcHal}{matrix} (lag, params, training, weights, base,
    project)\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
\method{calcHal}{matrix}(lag, params, training, weights, base, project)
}
\description{Method after Hallegatte}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

\alias{calcHal.matrix}

