\name{calcHal}
\alias{calcHal}
\title{Method after Hallegatte...}
\usage{calcHal(lag, params, training, weights, base, project)}
\description{Method after Hallegatte}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}}
\examples{ccafs_params <- dissimilarity(x, z, )}
