\name{cropInteractive}

\alias{cropInteractive}

\title{Crop raster to a region of intrest...}

\usage{
cropInteractive(r)
}

\description{
Crop raster to a region of intrest
}

\value{raster
}

\arguments{
  \item{rast}{RasterLayer or RasterStack}
}

\examples{
#assuming rs is a Raster* object
#rs.c <- cropInteractive(rs)
#then a window will pop up and you 
#will be able to select the cropping region 
#interactively
}

\alias{cropInteractive}


