\name{loadDataGRASS}

\alias{loadDataGRASS}

\title{Extract reference values from rasters using GRASS...}

\usage{
loadDataGRASS(params)
}

\description{Extract reference values from rasters using GRASS}

\value{
A number of \code{RasterLayer}s
}

\arguments{
\item{params}{an object of the class \code{list}, created using \code{createParameters()}}
}

\examples{

}

\alias{loadDataGRASS}



