\name{loadGridsFromFiles}
\alias{loadGridsFromFiles}
\title{Load seperate grids...}
\usage{loadGridsFromFiles(path, params)}
\description{Load seperate grids}
\details{This functions seperate grids and returns a raster stack}
\keyword{manip}
\value{an object of class RasterStack}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{path}{path to the file}}
