\name{madMPoints}

\alias{madMPoints}

\title{Method after Hallegatte...}

\usage{
madMPoints(ref.t, poi.t)
}

\description{
Method after Hallegatte. Users do not need to worry a lot about this function, as it is wrapped into \code{dissimilarity()} 
and \code{dissimilarityPoints()}
}

\arguments{
  \item{ref.t}{vector with values of vars at reference point}
  \item{poi.t}{matrix with values of vars at target points}
}

\examples{
#Nothing so far
}

\alias{madMPoints}

