\name{makeRef.AnalogueWeights}
\alias{makeRef.AnalogueWeights}
\title{Extract reference from weights grid...}
\usage{makeRef.AnalogueWeights(data, params)makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
makeRef.AnalogueWeights(data, params)
}
\description{Extract reference from weights grid}
\value{an object of AnalogueResults\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


\code{makeRef.AnalogueWeights}: an object of AnalogueResults


}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{weights}{a list with all the weights}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

\alias{makeRef.AnalogueWeights}

