\name{makeRef.ATraining}
\alias{makeRef.ATraining}
\title{Extract reference from data grid...}
\usage{makeRef.ATraining(data, params)makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
makeRef.ATraining(data, params)
}
\description{Extract reference from data grid}
\value{an object of AnalogueResults\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


\code{makeRef.ATraining}: an object of AnalogueResults


}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{weights}{a list with all the weights}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

\alias{makeRef.ATraining}

