\name{mrdMPoints}

\alias{mrdMPoints}

\title{Method after Hallegatte...}

\usage{
mrdMPoints(ref.t, poi.t)
}

\description{
Method after Hallegatte. Users do not need to worry a lot about this function, as it is wrapped into \code{dissimilarity()} 
and \code{dissimilarityPoints()}
}

\arguments{
  \item{ref.t}{vector with values of vars at reference point}
  \item{poi.t}{matrix with values of vars at target points}
}

\examples{
#Nothing at the moment
}

\alias{mrdMPoints}

