\name{radMPoints}
\alias{radMPoints}
\title{Method after Hallegatte...}
\usage{radMPoints(ref.t, poi.t)
radMPoints(ref.t, poi.t)
radMPoints(ref.t, poi.t)
radMPoints(ref.t, poi.t)
radMPoints(ref.t, poi.t)
}
\description{Method after Hallegatte}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{...}{more to come}
}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}

\alias{radMPoints}

\alias{radMPoints}

\alias{radMPoints}

\alias{radMPoints}

