\name{ref_vals}
\alias{ref_vals}
\title{
Extraction of values for reference site
}
\description{
Internal function. Used within \code{\link{calc_similarity}}
}
\usage{
ref_vals(params)
}
\arguments{
  \item{params}{List of parameters, created with function \code{\link{createParameters}}}
}
\value{
Vector of values extracted from the raster.
}
\seealso{
\code{\link{createParameters}},\code{\link{calc_similarity}}
}
\examples{
#no examples needed here
}
