\name{rota_serie}
\alias{rota_serie}
\title{
Rotate climate data
}
\description{
Internal function. Used within \code{\link{calc_similarity}}
}
\usage{
rota_serie(rotation, sit)
}
\arguments{
  \item{rotation}{Numeric value indicating the rotation point (month)}
  \item{sit}{Numeric vector of data to rotate over rotation point}
}
\seealso{
\code{\link{createParameters}},\code{\link{calc_similarity}}
}
\examples{
#no examples needed
}
