\name{applyThreshold}
\alias{applyThreshold}
\title{Apply thresholding to araster...}
\usage{applyThreshold(results, range=NA, best=NA)applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
applyThreshold(results, range=NA, best=NA)
}
\description{Apply thresholding to araster}
\value{RasterLayer\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


\code{applyThreshold}: RasterLayer


}
\arguments{\item{r}{RasterLayer or RasterStack}
\item{range}{if threshold should be applied to a range}
\item{best}{probablity values at which threshold should be cut}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

\alias{applyThreshold}

