\name{createParameters}
\alias{createParameters}
\title{Initiate parameters for analogue dissimilarity analysis.}
\usage{createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
createParameters(x=10, y=48, to=NA, method="ccafs", hal.rad=c(NA, 0.15), hal.mad=c(NA,
    0.3), hal.mrd=c(1, NA), hal.ncond=2, z=2, gcms=c("current"),
    vars=c("tmean", "prec"), weights=c("dtr", 1), ndivisions=12,
    climate.data=".", ext="asc", direction="backwd",
    growing.season=1:12, across.year=T, keep.lag=F, normalise=F)
}
\description{Initiate parameters for analogue dissimilarity analysis.}
\details{This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


\code{createParameters}: This function is the first step too find analogue climate sites.
Here all parameters for the later analysis are set. The initial 
parameters can be changed later. Make sure clmate data are name properly.


}
\keyword{manip}
\value{an object of AnalogueParameters\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


\code{createParameters}: an object of AnalogueParameters


}
\arguments{\item{x}{coordinate of site under analysis}
\item{y}{coodrinate of site under analysis}
\item{to}{coordinates for which dissimilarity should be caluclated as matrix with. A \code{matrix} 
with two columns: x and y. If is set to \code{NA} dissimilarities are calculated for the whole grid.}
\item{method}{select the method that is going to be use, either ccafs or hallegate}
\item{hal_rad}{maximum tolerable relative annual difference for hellegate method. A vector of \code{length(vars)}, 
if thre is a var that should not be considered, substitute the number with \code{NA}.}
\item{hal_mad}{maximum tollerable difference in annual precipitation. A vector of \code{length(vars)}, 
if thre is a var that should not be considered, substitute the number with \code{NA}.}
\item{hal_mrd}{mean absolute difference between months difference. A vector of \code{length(vars)}, 
if thre is a var that should not be considered, substitute the number with \code{NA}.}
\item{hal_ncond}{specify how many of the above condition need to be true}
\item{screnario}{specify the SRES (e.g. A1B, A2..)}
\item{gcms}{which gcm model should be used (e.g. bccr_bcm2_0,cccma_cgcm3_1_t47)}
\item{year}{specify the year for which the dissimilarity should be calculated}
\item{vars}{specify which variables are being used in the calculation (e.g. tmean, precipitation)}
\item{weights}{specify how the variables are weighted. This can either be a \code{rasterLayer}
or a single number. Provide a list with the name of the raster or a value for weighting.}
\item{ndivision}{define how many devision the variables have per year (e.g. for monthly data use 12).}
\item{climate_data}{define directory where the climate data is located}
\item{ext}{Extension of raster format (must be supported by gdal)}
\item{direction}{which direction should the dissimilarity be calculated. Available 
options are forward, backward and current.}
\item{growing_season}{define the growing season of the crop that is to be modelled.}
\item{accross_year}{should the analogue method be looked accross years (i.e. should time lag be 
used to account for differences in seasons}
\item{normalise}{should the rasater be normalised to a mean of 0 and sd of 1 when being loaded.}
\item{keep_lag}{specify whether or not grids for each lag should be kept}
\item{...}{expressions evaluated in the context of \code{df} and 
then fed to \code{\link{order}}}
\item{hal.rad}{maximum tolerable relative annual difference for hellegate method. A vector of \code{length(vars)}, 
if variables should not be considered, substitute the number with \code{NA}.}
\item{hal.mad}{maximum tollerable difference in annual precipitation. A vector of \code{length(vars)}, 
if variables should not be considered, substitute the number with \code{NA}.}
\item{hal.mrd}{mean absolute difference between months difference. A vector of \code{length(vars)}, 
if variables should not be considered, substitute the number with \code{NA}.}
\item{hal.ncond}{specify how many of the above condition need to be true in order that a sites is similiar.}
\item{accross.year}{should the analogue method be looked accross years (i.e. should time lag be 
used to account for differences in seasons}
\item{keepllag}{specify whether or not grids for each lag should be kept}
\item{z}{exponent for ccafs dissimilarity, default is 2.}
}
\examples{ccafs_params <- createParams(x, z, )ccafs_params <- createParams(x, z, )ccafs_params <- createParams(x, z, )ccafs_params <- createParams(x, z, )ccafs_params <- createParams(x, z, )}

\alias{createParameters}
\keyword{manip}

\alias{createParameters}
\keyword{manip}

\alias{createParameters}
\keyword{manip}

\alias{createParameters}
\keyword{manip}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

\alias{createParameters}

