\name{cropInteractive}
\alias{cropInteractive}
\title{Crop raster to a region of intrest...}
\usage{cropInteractive(r)cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
cropInteractive(r)
}
\description{Crop raster to a region of intrest}
\value{raster\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


\code{cropInteractive}: raster


}
\arguments{\item{rast}{RasterLayer or RasterStack}

}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

\alias{cropInteractive}

