\name{dissimilarity}
\alias{dissimilarity}
\title{Calculate dissimilarity...}
\usage{dissimilarity(params, training, weights)dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
dissimilarity(params, training, weights)
}
\description{Calculate dissimilarity}
\details{This functions calculate dissimilarities\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


\code{dissimilarity}: This functions calculate dissimilarities


}
\keyword{manip}
\value{an object of AnalogueResults\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


\code{dissimilarity}: an object of AnalogueResults


}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{training}{an object of the class AnalogueTraining}
\item{weights}{an object of the class AnalogueWeiths}}
\examples{ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )ccafs_params <- dissimilarity(x, z, )}
\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

\alias{dissimilarity}
\keyword{manip}

