\name{loadGridsFromFiles}
\alias{loadGridsFromFiles}
\title{Load seperate grids...}
\usage{loadGridsFromFiles(path, params)loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
loadGridsFromFiles(path, params)
}
\description{Load seperate grids}
\details{This functions seperate grids and returns a raster stack\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


\code{loadGridsFromFiles}: This functions seperate grids and returns a raster stack


}
\keyword{manip}
\value{an object of class RasterStack\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


\code{loadGridsFromFiles}: an object of class RasterStack


}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{path}{path to the file}}
\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

\alias{loadGridsFromFiles}
\keyword{manip}

