\name{radMPoints}

\alias{radMPoints}

\title{Method after Hallegatte...}

\usage{
radMPoints(ref.t, poi.t)
}

\description{Method after Hallegatte}

\arguments{
  \item{ref.t}{vector with values of vars at reference point}
  \item{poi.t}{matrix with values of vars at target points}
}

\examples{
#Nothing at the moment
}

\alias{radMPoints}

