\name{report}
\alias{report}
\title{Create a report...}
\usage{report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results)), pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
report(models=list(list(params, training, results, r.lab, m.lab)),
    pdf.name="test.pdf")
}
\description{Create a report}
\value{prints a report and save it as pdf\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


\code{report}: prints a report and save it as pdf


}
\arguments{\item{params}{an object of the class AnalogueParameters}
\item{r}{a \code{list} of rasters}
\item{r.lab}{\code{vector} of labels corresponding to \code{r}}
\item{pdf.name}{path and name to output pdf}
}
\examples{report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")report(ccafs.params, list(results[[1]]), "Test", "test.pdf")}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

\alias{report}

